+++ src/apis/carItems.ts
@@ -1,4 +1,3 @@
-import { CartItem } from '../types/type';
 import { END_POINTS } from './config';
 import response from './response';
