+++ src/hooks/useFetchCartItemList.ts
@@ -2,10 +2,9 @@ import { useQuery } from '@tanstack/react-query';
 import { fetchCartItemList } from '../apis/carItems';
 import { QUERY_KEYS } from '../apis/config';
 
-function useCartItemList() {
+export default function useFetchCartItemList() {
   return useQuery({
     queryKey: [QUERY_KEYS.CART_ITMES],
     queryFn: fetchCartItemList,
   });
 }
-export default useCartItemList;
